+++ src-docs/src/views/tables/custom/index.js
@@ -1,0 +1,1 @@
+export { section } from './custom_section';