+++ src-docs/src/views/text/text.js
@@ -14,6 +14,15 @@ export default () => (
         yellow sun.
       </p>
 
+      <blockquote>
+        <p>
+          I&apos;ve seen things you people wouldn&apos;t believe.
+          Attack ships on fire off the shoulder of Orion.
+          I watched C-beams glitter in the dark near the Tannhäuser Gate.
+          All those moments will be lost in time, like tears in rain. Time to die.
+        </p>
+      </blockquote>
+
       <p>
         Orbiting this at a distance of roughly ninety-two million miles
         is an utterly insignificant little blue green planet whose ape-