+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 - Added horizontal steps component ([#375](https://github.com/elastic/eui/pull/375))
 - Changed look and feel of pagination. Added compressed prop for smaller footprint pagination. ([#380](https://github.com/elastic/eui/pull/380))
 
+**Breaking changes**
+
+- Adds `EuiBasicTable` as an opinionated, high level component for constructing tables. Its addition deprecates `EuiTableOfRecords` which is still avaiable, but now marked for removal. ([#377](https://github.com/elastic/eui/pull/377))
+
 # [`0.0.20`](https://github.com/elastic/eui/tree/v0.0.20)
 
 - Rename class from `euiFlexGroup--alignItemsStart` to `euiFlexGroup--alignItemsFlexStart`